Rewrite CopyToClipboard as an ES Module
We want to migrate all of our legacy Javascript
files to ESM, so this is the latest in the series
of conversions.

Rewrite follows the same conventions we previously
used for FocusBanner, Collapsible checkboxes,
ColourPreview, FileUpload, Autofocus and
PreviewPane.

For copy to clipboard we switched to the Clipboard
API as it is prefered to the deprecated
`document.execCommand()` method for accessing the
clipboard.

Added bonus: removes the use of jQuery as well