Add more logic to template list visibility
There are some items that are hidden by default with CSS and which we need to show if they match the search query.

Whiel we do the visibility toggling by applying the `hidden` attribute to the each item, we cannot use the same attribute of the hiden-by-default items, as the checkbox macro from DS does not accomodate for that, hence they're hidden by css class.

This adds a bit more logic to the JS to apply a new class to those items when they match the search query and use CSS to show them, as well as remove the same class when they don't match or the search query is empty (default state)